Added links and tech entries
Added in all the technology entries, semi responsive things get pushed
around when on desktop.
Removed hyperlinks in header, was initially planning on three pages but
moved to a single one for convenience.
>Pages will get added as more things will be added.
Goals for next commit:
Fix the background repeating on mobile
Add in info for the rest of the tech section and fix desktop
responsiveness
Add a footer